pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "A voice-dictated diary that transcribes and stores locally"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "openai-whisper>=20241104",
    "sounddevice>=0.4.0",
    "scipy>=1.11.0",
    "ollama>=0.3.0",
]

[project.scripts]
diary = "diary.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
